Add block.difficulty and a-z file sorting

+++ src/models.rs
@@ -3,7 +3,7 @@ pub struct Block {
     pub block_hash: String,
     pub height: i32,
     pub time: chrono::NaiveDateTime,
-    pub difficulty: i64,
+    pub difficulty: f64,
     pub merkle_root: String,
     pub nonce: i64,
     pub size: i32,
